+++ R/feed_post_thread.R
@@ -6,6 +6,7 @@
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
+#' @param clean `r template_var_clean()`
 #'
 #' @concept feed
 #'
@@ -22,7 +23,7 @@
 #' bs_get_post_thread('at://did:plc:ic6zqvuw5ulmfpjiwnhsr2ns/app.bsky.feed.post/3k7qmjev5lr2s')
 bs_get_post_thread <- function(uri, depth = NULL, parent_height = NULL,
                               user = get_bluesky_user(), pass = get_bluesky_pass(),
-                              auth = bs_auth(user, pass)) {
+                              auth = bs_auth(user, pass), clean = TRUE) {
   if (missing(uri)) {
     cli::cli_abort('{.arg uri} must list at least one uri.')
   }
@@ -58,6 +59,10 @@ bs_get_post_thread <- function(uri, depth = NULL, parent_height = NULL,
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
+
+  if (!clean) {
+    return(resp)
+  }
 
   out <- resp |>
     purrr::pluck('thread') |>